Move tfsec formatter to conform to architectural standards
