+++ internal/services/formatters/hcl/tfsec/config.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package hcl
+package tfsec
 
 const CMD = `
 			{{WORK_DIR}}
