+++ internal/services/formatters/hcl/tfsec/entities/location.go
