+++ internal/services/formatters/hcl/tfsec/entities/vulnerabilities.go
